requirements.txt
streamlit
google.generativeai
PIL
pdf2image
os